requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-cuser==2017.3.16
django-filter==23.3
django-jazzmin==2.6.0
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
modeltranslation==0.25
packaging==23.1
Pillow==10.0.1
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
